Add u0

+++ hw2/include/function.h
@@ -8,10 +8,10 @@ class Function{
 private:
     double calculate_at();
 public:
-    unsigned int Lx, Ly, Lz;
+    double Lx, Ly, Lz;
     double at;
     Function(ConfigParser *parser);
-    Function(unsigned int Lx, unsigned int Ly, unsigned int Lz);
+    Function(double Lx, double Ly, double Lz);
     double operator ()(double x, double y, double z, double t);
     double phi(double x, double y, double z);
 };
